Moved Attributes to core (no dependencies)
