+++ FileHelpers/Core/RecordInfo.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Text.RegularExpressions;
+using FileHelpers.Core;
 using FileHelpers.Events;
 using FileHelpers.Helpers;
