+++ FileHelpers/Fields/FieldBase.cs
@@ -5,6 +5,7 @@ using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Text;
 using FileHelpers.Converters;
+using FileHelpers.Core;
 using FileHelpers.Helpers;
 using FileHelpers.Options;
